+++ CHANGELOG.md
@@ -11,6 +11,7 @@ SPDX-License-Identifier: HEESIL-1.0
 Added:
 - Mount actions listed in the 'actions' tab of a configuration now have the "edit" option, which redirect to the edit form of that mount action   
 - Mounts can have explicit coordinates
+- Shortcut to add a value for a parameter of device, platform or configuration 
 
 Fixed:
 - Adjusted the ordering of the elements in sensorML